Added original_sweep_start to the sweep start
AutoRaster at this point assumes that the sample is centered, but
orthogonal. So we take the current omega rotate it by 90 to be face on,
then we add the sweep start to this value. Usually the
original_sweep_start = 0 but when its -45 it will adjust accordingly.
This will allow any other angle that the scientists come up with